moved employe by manager function to managers route
