+++ src/routes/api/employees.js
@@ -4,7 +4,6 @@ const {
   addEmployee,
   removeEmployee,
   viewEmployeesByDepartment,
-  viewEmployeesByManager,
   updateEmployee,
 } = require("../../controllers/data/employees");
 
@@ -15,6 +14,5 @@ router.post("/", addEmployee);
 router.delete("/:id", removeEmployee);
 router.put("/:id", updateEmployee);
 router.get("/department/:id", viewEmployeesByDepartment);
-router.get("/manager/:id", viewEmployeesByManager);
 
 module.exports = router;
