+++ src/routes/api/managers.js
@@ -1,8 +1,12 @@
 const { Router } = require("express");
-const { getManagers } = require("../../controllers/data/managers");
+const {
+  getManagers,
+  viewEmployeesByManager,
+} = require("../../controllers/data/managers");
 
 const router = Router();
 
 router.get("/", getManagers);
+router.get("/:id", viewEmployeesByManager);
 
 module.exports = router;
